fastapi-incidentes/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
python-multipart